:art: :recycle: Move query-param related modules to one place
