+++ src/api/controllers/association.ts
@@ -2,7 +2,7 @@ import _ from 'lodash'
 import { Request, Response } from 'express'
 import { Controller } from '../../base/controller'
 import AssociationService from '../../services/association'
-import { resolveOptions } from '../common-query-params'
+import { resolveOptions } from '../params/common-query-params'
 import { ApiError } from '../../api/error/api-error'
 import { ApiErrorCode } from '../error/api-error-codes'
 import { instanceToPlain } from 'class-transformer'
