+++ src/services/member.ts
@@ -7,7 +7,7 @@ import { Service } from '../base/service'
 import { MemberItemsDto } from '../dto/member-items'
 import { MemberQueryOptions, MemberRepository } from '../repositories/member'
 import { ClientInfo } from '../utils/jwt'
-import { CommonQueryOptions } from '../api/common-query-params'
+import { CommonQueryOptions } from '../api/params/common-query-params'
 import { MemberDto } from '../dto/member'
 import { MemberInviteDto } from '../dto/member-invite'
 import { MailService } from './mail'
